mon: Elector: bootstrap and reprobe if peer knows of other mons we don't
If for some reason we got into an election with knowledge of only a
portion of all existing monitors, and if we happen to be the leader,
there's a chance that not only will we never get to know all monitors in
the cluster, but the fact that other monitors do will cause havoc -- e.g.,
elections being triggered to include the other monitors, while we ignore
them because we don't really have them in our monmap.

This patch should work around that by forcing the monitor to bootstrap and
probe other peers if upon an election some other monitor tells us about
some peer not in our monmap (rather, with a blank ip) but present in our
extra_probe_peers set -- meaning they probed us at some point, but we did
not probe back (most likely because we were already electing or elected).

Fixes: 7212

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>